+++ talker_agent/install_agent.sh
@@ -23,7 +23,7 @@ fi
 echo "$TALKER_AGENT_VERSION" > version
 
 # Add redis py dependency
-REDIS_COMMIT_ID=ad84781ea260be0a1ca4bf6768959b50e8835a6b
-curl -fL https://github.com/weka-io/redis-py/archive/"$REDIS_COMMIT_ID".tar.gz | tar -xz
-mv redis-py-"$REDIS_COMMIT_ID"/redis .
-rm -rf redis-py-"$REDIS_COMMIT_ID"
+TAG=3.3.11
+curl -fL https://github.com/andymccurdy/redis-py/archive/"$TAG".tar.gz | tar -xz
+mv redis-py-"$TAG"/redis .
+rm -rf redis-py-"$TAG"