+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Redis connection parameters in agent config file
+
 ### Changed
+- raise client's Redis socket timeout
 - using .ini file for agent configuration
 
 ## [1.8.4] - 2019-09-22